% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bars}
\alias{bars}
\title{Vertical Bars}
\format{A matrix, 20 rows by 20 columns}
\usage{
bars
}
\description{
A Matrix of vertical bars, each column with a different value. Column 1
contains the value 1, up to column 20 with value 20.
}
\keyword{datasets}
