% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalculateFeatures.R
\name{calc_features}
\alias{calc_features}
\title{Calculate texture and first order statistics.}
\usage{
calc_features(image, features = c("first order", "glcm", "glrlm", "glszm",
  "mglszm"), n_grey = 32, d = 1, verbose = FALSE,
  max_run_length = min(dim(image)))
}
\arguments{
\item{image}{A numeric image matrix.}

\item{features}{A vector containing any of "first order", "glcm", "glrlm",
"glszm", and "mglszm". Inclusion of these strings in the vector will lead
to the calculation of the related image features.}

\item{n_grey}{an integer value, the number of grey levels the image should
be quantized into.}

\item{d}{an integer value, the distance between the current pixel, and the
pixel to which it is compared. To be passed to \code{glcm}.}

\item{verbose}{Logical, a warning is given when the user
supplies more grey values than exist in the image. Setting this value to FALSE
will suppress this warning.}

\item{max_run_length}{An integer value, the default is the maximum possible
run length. Setting it to a smaller value truncates the output. Desirable
in cases where the matrix is extremely sparse, for example when
there are few long runs. To be passed to \code{glrlm}.}
}
\value{
A data frame with a single observation. The columns of the dataframe
  correspond to the calculated features.
}
\description{
\code{calc_features} calculates the first order and textural features for a given matrix
}
\examples{
calc_features(hallbey)
calc_features(psf, n_grey=10)
}
\seealso{
\code{\link{glcm}}
  \code{\link{glrlm}}
  \code{\link{glszm}}
  \code{\link{mglszm}}
}

