% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageQuantize.R
\name{discretizeImage2}
\alias{discretizeImage2}
\title{Image Discretization.}
\usage{
discretizeImage2(image, n_grey = 32)
}
\arguments{
\item{image}{A numeric image matrix.}

\item{n_grey}{The grey levels the output image will have}
}
\description{
#' \code{discretizeImage2} Scales the grey values of an image into a specified number of values.
}
\details{
Not currently used. Different methods of discretizing the image will be explored in the future.
}
