% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FirstOrder.R
\name{first_order_features}
\alias{calc_energy}
\alias{calc_entropy}
\alias{calc_kurtosisOptimized}
\alias{calc_meanDeviation}
\alias{calc_skewness}
\alias{calc_uniformity}
\alias{first_order_features}
\title{First order features}
\usage{
calc_energy(image, dims = NULL)

calc_entropy(image, base = 2, nbins = length(image))

calc_kurtosisOptimized(image)

calc_meanDeviation(image)

calc_skewness(image, na.rm = FALSE)

calc_uniformity(image, nbins = length(image))
}
\arguments{
\item{image}{Numeric image matrix.}
}
\description{
First order features
}
\section{Functions}{
\itemize{
\item \code{calc_energy}: Energy (ASM)

\item \code{calc_entropy}: Entropy

\item \code{calc_kurtosisOptimized}: Energy (ASM)

\item \code{calc_meanDeviation}: Mean Deviation

\item \code{calc_skewness}: Skewness

\item \code{calc_uniformity}: Uniformity
}}
\references{
\url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0102107#s5}
}

