% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GLCM.R
\name{glcm}
\alias{glcm}
\title{Gray level co-occurrence matrix.}
\usage{
glcm(image, angle = "0", d = 1, n_grey = length(unique(c(image))),
  normalize = TRUE, ...)
}
\arguments{
\item{angle}{One of "0", "45", "90" or "135", the pixel to which the
current pixel is compared.}

\item{d}{an integer value, the distance between the current pixel, and the
pixel to which it is compared.}

\item{n_grey}{an integer value, the number of grey levels the image should
be quantized into.}

\item{normalize}{Logical value, if TRUE the matrix will be normalized such that
the sum of it's components is 1.}

\item{...}{Can be given verbose=FALSE to suppress output from the n_grey conversion.}
}
\value{
a matrix of dimension n_grey by n_grey, the GLCM.
  See \url{http://www.fp.ucalgary.ca/mhallbey/tutorial.htm} for details.
}
\description{
\code{glcm} returns a gray level co-occurrence matrix for a given matrix.
}
\details{
This function can be used alone, or can be executed and it's textural
features automatically calculated using \code{calc_features}
}
\examples{
hallbey
glcm(hallbey)
glcm(hallbey, angle="90") #vertical GLCM
}

