% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GLRLM.R
\name{glrlm}
\alias{glrlm}
\title{Gray level run length matrix.}
\usage{
glrlm(image, angle = "0", n_grey = 32, max_run_length = min(dim(image)),
  ...)
}
\arguments{
\item{image}{A numeric image matrix.}

\item{angle}{One of "0", "45", "90" or "135", the direction the run is calculated.}

\item{n_grey}{an integer value, the number of grey levels the image should
be quantized into.}

\item{max_run_length}{An integer value, the default is the maximum possible
run length. Setting it to a smaller value truncates the output. Desirable
in cases where the matrix is extremely sparse, for example when
there are few long runs.}

\item{...}{Can be given verbose=FALSE to suppress output from the n_grey conversion.}
}
\value{
a matrix of dimension n_grey by run length, the GLRLM. The column
  names represent the length of the run, and row names represent
  grey values in the image.
  See Galloway 1974
  (\url{http://www.sciencedirect.com/science/article/pii/S0146664X75800086}) for details.
}
\description{
\code{glrlm} returns a gray level run length matrix for a given matrix.
}
\details{
This function can be used alone, or can be executed and it's textural
features automatically calculated using \code{calc_features}.
}
\examples{
hallbey
glrlm(hallbey)
glrlm(hallbey, angle="90")
}

