% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLRLMFeatures.R
\name{glrlm_features}
\alias{glrlm_features}
\alias{glrlm_GLN}
\alias{glrlm_HGLRE}
\alias{glrlm_LRE}
\alias{glrlm_LRHGLE}
\alias{glrlm_LRLGLE}
\alias{glrlm_LGLRE}
\alias{glrlm_RLN}
\alias{glrlm_RP}
\alias{glrlm_SRE}
\alias{glrlm_SRHGLE}
\alias{glrlm_SRLGLE}
\title{GLRLM Features}
\usage{
glrlm_GLN(glrlm)

glrlm_HGLRE(glrlm)

glrlm_LRE(glrlm)

glrlm_LRHGLE(glrlm)

glrlm_LRLGLE(glrlm)

glrlm_LGLRE(glrlm)

glrlm_RLN(glrlm)

glrlm_RP(glrlm)

glrlm_SRE(glrlm)

glrlm_SRHGLE(glrlm)

glrlm_SRLGLE(glrlm)
}
\arguments{
\item{glrlm}{A matrix of class "glrlm" produced by \code{glrlm}.}
}
\description{
GLRLM Features
}
\section{Functions}{
\itemize{
\item \code{glrlm_GLN}: Grey level non-uniformity

\item \code{glrlm_HGLRE}: High Gray level run emphasis

\item \code{glrlm_LRE}: Long Run Emphasis

\item \code{glrlm_LRHGLE}: Long run high gray level emphasis

\item \code{glrlm_LRLGLE}: Long Run Low Gray Level Emphasis

\item \code{glrlm_LGLRE}: Low gray level run emphasis

\item \code{glrlm_RLN}: Run length non-uniformity

\item \code{glrlm_RP}: Run Percentage

\item \code{glrlm_SRE}: Short run emphasis

\item \code{glrlm_SRHGLE}: rt run high gray level emphasis

\item \code{glrlm_SRLGLE}: Short run low grey emphasis
}}

\references{
\url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0102107#s5}
}
