% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLSZMFeatures.R
\name{glszm_features}
\alias{glszm_features}
\alias{glszm_SAE}
\alias{glszm_LAE}
\alias{glszm_IV}
\alias{glszm_SZV}
\alias{glszm_ZP}
\alias{glszm_LIE}
\alias{glszm_HIE}
\alias{glszm_LISAE}
\alias{glszm_HISAE}
\alias{glszm_LILAE}
\alias{glszm_HILAE}
\title{GLSZM Features}
\usage{
glszm_SAE(glszm)

glszm_LAE(glszm)

glszm_IV(glszm)

glszm_SZV(glszm)

glszm_ZP(glszm)

glszm_LIE(glszm)

glszm_HIE(glszm)

glszm_LISAE(glszm)

glszm_HISAE(glszm)

glszm_LILAE(glszm)

glszm_HILAE(glszm)
}
\arguments{
\item{glszm}{A matrix of class "glszm" produced by \code{glszm}.}
}
\description{
GLSZM Features
}
\section{Functions}{
\itemize{
\item \code{glszm_SAE}: Small Area Emphasis

\item \code{glszm_LAE}: Large Area Emphasis

\item \code{glszm_IV}: Intensity Variability

\item \code{glszm_SZV}: Size Zone Variability

\item \code{glszm_ZP}: Zone percentage

\item \code{glszm_LIE}: Low intensity emphasis

\item \code{glszm_HIE}: High intensity emphasis

\item \code{glszm_LISAE}: Low intensity small area emphasis

\item \code{glszm_HISAE}: High intensity small area emphasis

\item \code{glszm_LILAE}: Low intensity large area emphasis

\item \code{glszm_HILAE}: High intensity Large area emphasis
}}

\references{
\url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0102107#s5}
}
