% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallbey}
\alias{hallbey}
\title{Hall Bey}
\format{A matrix, 4 rows by 4 columns}
\usage{
hallbey
}
\description{
A Matrix used as examples in the hallbey explanation of glcms.
}
\keyword{datasets}
