% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psf}
\alias{psf}
\title{Point Spread Function}
\format{A matrix, 50 rows by 50 columns}
\usage{
psf
}
\description{
A Matrix of a point spread function.
Values are smallest in the middle, and increase in a radial fashion.
}
\keyword{datasets}
