% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiomics.R
\docType{package}
\name{radiomics}
\alias{radiomics}
\alias{radiomics-package}
\title{radiomics: A texture analysis toolbox for image classification}
\description{
radiomics provides a several new classes of matrices: GLCM (grey level co-occurrence matrix),
GLRLM (grey level run-length matrix), GLSZM (grey level size-zone matrix), and the MGLSZM 
(multiple GLSZM).
}
\details{
To learn more about radiomics and texture matrices, start with the vignettes:
\code{browseVignettes(package = "radiomics")}
}
