% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumor}
\alias{tumor}
\title{Brain Tumor Slice}
\format{A matrix, 47 rows by 46 columns}
\usage{
tumor
}
\description{
A Matrix of a single image slice of a tumor taken from an MRI image.
This slice was extracted from one of the sample data sets of 3DSlicer.
}
\keyword{datasets}
